refactor(api): move assessment-result repository
